fix: alpha for legacy hsl
