+++ crates/swc_css_compat/src/compiler/mod.rs
@@ -12,8 +12,8 @@ mod color_alpha_parameter;
 mod color_hex_alpha;
 mod color_space_separated_parameters;
 mod custom_media;
+mod legacy_rgb_and_hsl;
 mod media_query_ranges;
-mod percentage_in_parameters;
 mod utils;
 
 /// Compiles a modern CSS file to a CSS file which works with old browsers.
@@ -155,8 +155,8 @@ impl VisitMut for Compiler {
             self.process_color_alpha_parameter(n);
         }
 
-        if process.contains(Features::COLOR_PERCENTAGE_IN_PARAMETERS) {
-            self.process_percentage_in_parameters(n);
+        if process.contains(Features::COLOR_LEGACY_RGB_AND_HSL) {
+            self.process_rgb_and_hsl(n);
         }
     }
 }
