+++ crates/swc_css_compat/tests/fixture.rs
@@ -159,8 +159,8 @@ fn test_color_space_separated_function_notation(input: PathBuf) {
 
         ss.visit_mut_with(&mut Compiler::new(Config {
             process: Features::COLOR_SPACE_SEPARATED_PARAMETERS
-                | Features::COLOR_PERCENTAGE_IN_PARAMETERS
-                | Features::COLOR_ALPHA_PARAMETER,
+                | Features::COLOR_ALPHA_PARAMETER
+                | Features::COLOR_LEGACY_RGB_AND_HSL,
         }));
 
         let s = print_stylesheet(&ss);
